Update the starred eps and add a crude script to sync with AntennaPod backups.
The script doesn't handle 'shortening' of episodes, nor does it handle that I
never 'backported' starring episodes fully into AntennaPod, so some are starred
only here and not there. There's not much benefit in doing that given here will
be the canonical data source, but it means the script doesn't fully match what's
here and there, there are some extras.